+++ app/(tabs)/vehicle/[vehicle_id]/index.tsx
@@ -1,67 +1,36 @@
-import React, { useEffect, useState } from 'react';
+import React from 'react';
 import { StyleSheet } from 'react-native';
-import { View, Text, FlatList } from '@app/components/elements';
+import { View, Text, FlatList, Ionicons } from '@app/components/elements';
 import RecordCard from '@app/components/cards/record';
 import { router, useLocalSearchParams } from 'expo-router';
-import { useStore } from 'tinybase/ui-react';
-import { schema, tables } from '@app/database/schema';
-import { createQueries } from 'tinybase';
-import CarNicknameInHeader from '@app/components/carHeader';
-import CallbackButton from '@app/components/callbackButton';
+import { useSliceRowIds } from 'tinybase/ui-react';
+import VehicleNicknameInHeader from '@app/components/hooks/vehicleHeader';
+import { OptionButtons } from '@app/components/elements/optionButtons';
 
 export default function Records(): React.ReactElement {
   const { vehicle_id } = useLocalSearchParams<{ vehicle_id: string }>();
-  const store = useStore();
-  const queries = createQueries(store);
 
-  const [records, setRecords] = useState({});
-
-  useEffect(() => {
-    const toRemove = queries.addResultTableListener('vehicleRecords', (queries) => {
-      setRecords(queries.getResultTable('vehicleRecords'));
-    });
-
-    queries.setQueryDefinition(
-      'vehicleRecords',
-      tables.maintenance_records,
-      ({ select, where }) => {
-        for (const column of Object.keys(schema.maintenance_records)) {
-          select(column);
-        }
-        where('car_id', vehicle_id);
-      },
-    );
-
-    return () => {
-      queries.delListener(toRemove);
-      queries.delQueryDefinition('vehicleRecords');
-    };
-  }, [vehicle_id]);
-
+  const recordIDs = useSliceRowIds('byVehicle', vehicle_id);
   return (
     <View style={ pageStyles.container }>
-      <CarNicknameInHeader />
+      <VehicleNicknameInHeader />
       <FlatList
-        data={ Object.keys(records).sort((a, b) => {
-          return records[a].date < records[b].date ? 1 : -1;
-        }).map((key) => {
-          return { ...records[key], id: key };
-        }) }
-        renderItem={({ item }) => <RecordCard key={ (item as { id: string }).id } record={ item } /> }
+        data={ recordIDs }
+        renderItem={({ item }) => <RecordCard key={ item } record_id={ item } /> }
         ListEmptyComponent={
           <Text style={ pageStyles.emptyText }>No records</Text>
         }
       />
-      <View style={ pageStyles.actionButtonSection }>
-        <CallbackButton
-          pressable={{ style: pageStyles.actionButton }}
-          text={{ style: pageStyles.actionButtonText }}
-          title="Add Maintenance"
-          onPress={() => {
-            router.push(`/vehicle/${vehicle_id}/record/add`);
-          }}
-        />
-      </View>
+      <OptionButtons
+        options={[
+          { label: 'Add Maintenance', key: 'add', icon: <Ionicons name="add-circle" size={20} /> },
+        ]}
+        onSelect={ (value, enable) => {
+          router.push(`/vehicle/${vehicle_id}/record/add`);
+          enable();
+        }}
+        highlightAll={true}
+      />
     </View>
   );
 }