+++ app/(tabs)/vehicle/[vehicle_id]/record/add.tsx
@@ -1,12 +1,12 @@
 import React from 'react';
 import { KeyboardAwareScrollView } from '@app/components/elements';
-import CarNicknameInHeader from '@app/components/carHeader';
+import VehicleNicknameInHeader from '@app/components/hooks/vehicleHeader';
 import RecordForm from '@app/components/forms/record';
 
 export default function Add(): React.ReactElement {
   return (
     <KeyboardAwareScrollView>
-      <CarNicknameInHeader />
+      <VehicleNicknameInHeader />
       <RecordForm />
     </KeyboardAwareScrollView>
   );