+++ components/hooks/vehicleHeader.tsx
@@ -3,10 +3,10 @@ import { useLocalSearchParams, useNavigation } from 'expo-router';
 import { useEffect } from 'react';
 import { useRow } from 'tinybase/ui-react';
 
-export default function CarNicknameInHeader() {
+export default function VehicleNicknameInHeader() {
   const { vehicle_id } = useLocalSearchParams<{ vehicle_id?: string }>();
   const navigation = useNavigation();
-  const row = useRow(tables.cars, vehicle_id ?? '');
+  const row = useRow(tables.vehicles, vehicle_id ?? '');
 
   useEffect(() => {
     if (vehicle_id) {