+++ ChangeLog.md
@@ -3,7 +3,7 @@
 ## 1.0.2
 
 - Removed KeyboardAvoidingView, as it was causing crashing bug (Discoverd by Google)
-- Added some common maintainance types to the database
+- Added some common maintenance types to the database
 
 ## 1.0.1
 
@@ -11,7 +11,7 @@ Initial Release!
 
 Functionality:
 
-- CRUD for cars and maintainances
+- CRUD for cars and maintenance records
 - Dark and Light mode
 - WatermelonDB for offline storage
 - Database stores distances in Miles, but setting for display of distances in Kilometers